Add BooksRepository.kt Update BookClient.kt, BookAdapter.kt, BookTopicFragment.kt, BookTopicViewModel.kt, and build.gradle
