+++ app/build.gradle
@@ -59,6 +59,4 @@ dependencies {
     implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0"
     implementation "androidx.fragment:fragment-ktx:1.3.2"
     kapt "com.android.databinding:compiler:4.1.1"
-    //debugging
-    implementation("com.squareup.okhttp3:logging-interceptor:4.9.1")
 }
