+++ app/src/main/java/com/dteti/bookapp/data/api/BookClient.kt
@@ -2,21 +2,14 @@ package com.dteti.bookapp.data.api
 
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
-import okhttp3.OkHttpClient
-import okhttp3.logging.HttpLoggingInterceptor
 
 object BookClient {
     private const val base_url = "https://www.googleapis.com/books/v1/"
-
-    private val logging = HttpLoggingInterceptor().setLevel(HttpLoggingInterceptor.Level.BODY)
-    private val okHttpClient = OkHttpClient.Builder()
-        .addInterceptor(logging)
 
     val instance : BooksApi by lazy {
         val retrofitBuilder = Retrofit.Builder()
             .baseUrl(base_url)
             .addConverterFactory(GsonConverterFactory.create())
-            .client(okHttpClient.build())
             .build()
 
         retrofitBuilder.create(BooksApi::class.java)
